Structured rendering 'through' the camera
This is mostly a 'minor' change, revolving around a structured
configuration of various parameters that influence how rendering the
entire scene progresses.

Moreover, scenes can now be rendered in different styles which control
the order in which pixels on the canvas are processed. As of now:

    - scanline and
    - hilbert-curve

orderings are supported.

Apart from this, it seems that with "-ffast-math" option llvm's code
for cube-test fails. Remove that option for now.

+++ src/io/canvas.hpp
@@ -21,6 +21,7 @@
 /// our includes
 #include "common/include/logging.h"
 #include "primitives/color.hpp"
+#include "primitives/ray.hpp"
 
 namespace raytracer
 {
@@ -84,7 +85,7 @@ namespace raytracer
                 /// ------------------------------------------------------------
                 /// MUTATORS
 
-                void write_pixel(size_t x, size_t y, color c)
+                void write_pixel(uint32_t x, uint32_t y, color c)
                 {
                         this->buf_[x + y * this->width_] = c;
                 }
